blog 4.21 complete - blog only deleted by user who added

+++ part4/blog/controllers/blogs.js
@@ -3,13 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -26,8 +19,7 @@ blogsRouter.post('/', async (request, response, next) => {
         error: 'title and url are required'
       })
     }
-
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
       return response.status(401).json({ error: 'token invalid' })
     }
@@ -77,6 +69,23 @@ blogsRouter.get('/:id', (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  if(!request.token) {
+    return response.status(401).json({ error: 'token invalid' }).end()
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog) {
+    return response.status(404).end()
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
